backend/requirements.txt
django
django-cors-headers
djangorestframework
pandas
scikit-learn==1.5.2
xgboost